Redesign filters on report
