+++ fabric/src/components/TextInput/index.tsx
@@ -11,6 +11,7 @@ export type TextInputProps = React.InputHTMLAttributes<HTMLInputElement> &
   InputUnitProps & {
     action?: React.ReactNode
     symbol?: React.ReactNode
+    row?: boolean
   }
 export type TextAreaInputProps = React.InputHTMLAttributes<HTMLTextAreaElement> &
   InputUnitProps & {
@@ -113,11 +114,12 @@ export function TextInputBox(
   props: Omit<TextInputProps, 'label' | 'secondaryLabel'> & {
     error?: boolean
     inputRef?: React.Ref<HTMLInputElement>
+    row?: boolean
   }
 ) {
-  const { error, disabled, action, symbol, inputRef, inputElement, ...inputProps } = props
+  const { error, disabled, action, symbol, inputRef, inputElement, row, ...inputProps } = props
   return (
-    <StyledInputBox alignItems="stretch" height="input">
+    <StyledInputBox hideBorder={!!row} alignItems="stretch" height="input">
       {inputElement ?? <StyledTextInput disabled={disabled} {...inputProps} id={useContextId()} ref={inputRef} />}
       {symbol && (
         <Flex alignSelf="center" pr={1}>
@@ -167,9 +169,10 @@ export function SearchInput({ label, secondaryLabel, disabled, errorMessage, id,
   )
 }
 
-export function DateInput({ label, secondaryLabel, disabled, errorMessage, id, ...inputProps }: TextInputProps) {
+export function DateInput({ label, secondaryLabel, disabled, errorMessage, id, row, ...inputProps }: TextInputProps) {
   const defaultId = React.useId()
   id ??= defaultId
+
   return (
     <InputUnit
       id={id}
@@ -177,12 +180,14 @@ export function DateInput({ label, secondaryLabel, disabled, errorMessage, id, .
       secondaryLabel={secondaryLabel}
       disabled={disabled}
       errorMessage={errorMessage}
+      row={row}
       inputElement={
         <TextInputBox
           type="date"
           disabled={disabled}
           error={!!errorMessage}
           required // hides the reset button in Firefox
+          row={row}
           {...inputProps}
         />
       }
